+++ src/errors.rs
@@ -1,7 +1,3 @@
-
-
-
-
 use std::{
     error::Error as StdError,
     fmt::{self, Display},
