+++ src/structs/mod.rs
@@ -1,0 +1,8 @@
+mod link;
+pub use link::Link;
+
+#[derive(Debug, Default)]
+pub struct RepostCount {
+    pub link: String,
+    pub count: u64,
+}
